feat(s1-28): active nav highlighting for social section
Extracts the nav links into SocialNavClient (usePathname) so the
active segment is visually highlighted. Exact + prefix match handles
/posts vs /posts/[id]. Layout imports the client component; server
auth gate unchanged.
